pyproject.toml
[tool.poetry]
name = "gnc-dashboard"
version = "0.1.0"
description = ""
authors = ["Tsmorz <tonysmoragiewicz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
plotly = "^5.23.0"
geopandas = "^1.0.1"
shapely = "^2.0.5"
pandas = "^2.2.2"
dash = "^2.17.1"
pyorbital = "^1.8.3"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
